Doc needs more formatting
Doc needs more formatting as it's currently not very clear. It could be a good idea to change up the categories, have them more cleanly delineated, and beyond that a bit more detail where it matters, and add an extra level of docs for the more complicated topics. but where a dev is likely to be expected to know things.